Thu, Jul 13, 2023, 5:35 PM -07:00

+++ assignemt/3.py
@@ -1,3 +1,3 @@
 PI=3.14
 r=int(input('Enter area of RDIOUS :'))
-float(print("Area of circula :",PI*r*r))
+print("Area of circula :",PI*r*r)
